# frontend/src/lib/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateEnrichmentStatus": {
        "properties": {
          "candidate_id": {
            "format": "uuid",
            "title": "Candidate Id",
            "type": "string"
          },
          "job": {
            "$ref": "#/components/schemas/JobStatusResponse"
          },
          "policy_version": {
            "minimum": 1.0,
            "title": "Policy Version",
            "type": "integer"
          }
        },
        "required": [
          "candidate_id",
          "policy_version",
          "job"
        ],
        "title": "CandidateEnrichmentStatus",
        "type": "object"
      },
...
      "EnrichmentJobResult": {
        "additionalProperties": false,
        "properties": {
          "acquisition_avoided": {
            "default": false,
            "title": "Acquisition Avoided",
            "type": "boolean"
          },
          "covered_aspects": {
            "items": {
              "type": "string"
            },
            "maxItems": 32,
            "title": "Covered Aspects",
            "type": "array"
          },
          "covered_count": {
            "maximum": 32.0,
            "minimum": 0.0,
            "title": "Covered Count",
            "type": "integer"
          },
          "limitations": {
            "items": {
              "$ref": "#/components/schemas/EnrichmentLimitation"
            },
            "maxItems": 10,
            "title": "Limitations",
            "type": "array"
          },
          "missing_aspects": {
            "items": {
              "type": "string"
            },
            "maxItems": 32,
            "title": "Missing Aspects",
            "type": "array"
          },
          "missing_count": {
            "maximum": 32.0,
            "minimum": 0.0,
            "title": "Missing Count",
            "type": "integer"
          },
          "outcome": {
            "enum": [
              "complete",
              "partial"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "policy_version": {
            "minimum": 1.0,
            "title": "Policy Version",
            "type": "integer"
          }
        },
        "required": [
          "outcome",
          "policy_version",
          "covered_aspects",
          "missing_aspects",
          "covered_count",
          "missing_count"
        ],
        "title": "EnrichmentJobResult",
        "type": "object"
      },
      "EnrichmentLimitation": {
        "enum": [
          "missing_required_aspects",
          "safety_evidence_rejected"
        ],
        "title": "EnrichmentLimitation",
        "type": "string"
      },
      "GardenDeleteResponse": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "GardenDeleteResponse",
        "type": "object"
      },
...
      "GardenPlantCreate": {
        "properties": {
          "confirmed_candidate_id": {
            "format": "uuid",
            "title": "Confirmed Candidate Id",
            "type": "string"
          },
          "custom_data": {
            "additionalProperties": true,
            "title": "Custom Data",
            "type": "object"
          },
          "image_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Path"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "required": [
          "confirmed_candidate_id"
        ],
        "title": "GardenPlantCreate",
        "type": "object"
      },
      "GardenPlantResponse": {
        "properties": {
          "active_reminders": {
            "default": 0,
            "title": "Active Reminders",
            "type": "integer"
          },
          "confirmed_candidate_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed Candidate Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "custom_data": {
            "additionalProperties": true,
            "title": "Custom Data",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Path"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "profile": {
            "$ref": "#/components/schemas/PlantProfileResponse"
          }
        },
        "required": [
          "id",
          "profile",
          "created_at"
        ],
        "title": "GardenPlantResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobFailureCategory": {
        "enum": [
          "invalid_payload",
          "unsupported_payload_version",
          "unknown_job_type",
          "database_transient",
          "provider_transient",
          "indexing_transient",
          "invariant_violation",
          "attempts_exhausted",
          "unexpected_error",
          "lease_expired",
          "lease_lost",
          "insufficient_evidence"
        ],
        "title": "JobFailureCategory",
        "type": "string"
      },
      "JobLimitation": {
        "enum": [
          "some_claims_failed",
          "indexing_deferred"
        ],
        "title": "JobLimitation",
        "type": "string"
      },
      "JobStatus": {
        "enum": [
          "pending",
          "processing",
          "complete",
          "partial",
          "failed"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "JobStatusResponse": {
        "properties": {
          "attempt_count": {
            "minimum": 0.0,
            "title": "Attempt Count",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_type": {
            "$ref": "#/components/schemas/JobType"
          },
          "last_error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReadJobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "max_attempts": {
            "minimum": 1.0,
            "title": "Max Attempts",
            "type": "integer"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReadJobResult"
              },
              {
                "$ref": "#/components/schemas/EnrichmentJobResult"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "job_type",
          "status",
          "attempt_count",
          "max_attempts",
          "created_at",
          "updated_at"
        ],
        "title": "JobStatusResponse",
        "type": "object"
      },
      "JobType": {
        "enum": [
          "ingest_validated_claims",
          "enrich_confirmed_plant"
        ],
        "title": "JobType",
        "type": "string"
      },
...
      "PlantProfileResponse": {
        "properties": {
          "aliases": {
            "items": {
              "$ref": "#/components/schemas/ProfileAlias"
            },
            "title": "Aliases",
            "type": "array"
          },
          "common_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Common Name"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "enrichment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateEnrichmentStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "scientific_name": {
            "title": "Scientific Name",
            "type": "string"
          },
          "sections": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Sections",
            "type": "object"
          },
          "selected_alias": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Alias"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/ProfileSource"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "id",
          "scientific_name",
          "confidence"
        ],
        "title": "PlantProfileResponse",
        "type": "object"
      },
      "ProfileAlias": {
        "properties": {
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProfileAlias",
        "type": "object"
      },
      "ProfileSource": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "title",
          "url",
          "domain",
          "confidence"
        ],
        "title": "ProfileSource",
        "type": "object"
      },
...
      "ReadJobError": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/JobFailureCategory"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "category"
        ],
        "title": "ReadJobError",
        "type": "object"
      },
      "ReadJobResult": {
        "properties": {
          "failed": {
            "default": 0,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Failed",
            "type": "integer"
          },
          "limitations": {
            "items": {
              "$ref": "#/components/schemas/JobLimitation"
            },
            "maxItems": 10,
            "title": "Limitations",
            "type": "array"
          },
          "partial": {
            "default": false,
            "title": "Partial",
            "type": "boolean"
          },
          "skipped": {
            "default": 0,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Skipped",
            "type": "integer"
          },
          "succeeded": {
            "default": 0,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Succeeded",
            "type": "integer"
          }
        },
        "title": "ReadJobResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Fotosintesis AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/garden": {
      "get": {
        "operationId": "list_garden_plants_garden_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "fotosintesis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fotosintesis Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GardenPlantResponse"
                  },
                  "title": "Response List Garden Plants Garden Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Garden Plants",
        "tags": [
          "plant-profile-garden"
        ]
      },
      "post": {
        "operationId": "save_garden_plant_garden_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "fotosintesis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fotosintesis Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GardenPlantCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GardenPlantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Garden Plant",
        "tags": [
          "plant-profile-garden"
        ]
      }
    },
    "/garden/{garden_id}": {
      "delete": {
        "operationId": "delete_garden_plant_garden__garden_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "garden_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Garden Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "confirm_reminders",
            "required": false,
            "schema": {
              "default": false,
              "title": "Confirm Reminders",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "fotosintesis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fotosintesis Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GardenDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Garden Plant",
        "tags": [
          "plant-profile-garden"
        ]
      },
      "get": {
        "operationId": "get_garden_plant_garden__garden_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "garden_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Garden Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "fotosintesis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fotosintesis Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GardenPlantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Garden Plant",
        "tags": [
          "plant-profile-garden"
        ]
      }
    },
...
    "/plant-profiles/{scientific_name}": {
      "get": {
        "operationId": "get_plant_profile_plant_profiles__scientific_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "scientific_name",
            "required": true,
            "schema": {
              "title": "Scientific Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "candidateId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidateid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          },
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "fotosintesis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fotosintesis Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Plant Profile",
        "tags": [
          "plant-profile-garden"
        ]
      }
    },
...
  }
}